Added table to display entered bill values

+++ database.py
@@ -1,7 +1,7 @@
 from google.appengine.ext import ndb
 
 class DatabaseBill(ndb.Model):
-    util_qty = ndb.IntegerProperty()
-    util_cost = ndb.IntegerProperty()
+    util_qty = ndb.FloatProperty()
+    util_cost = ndb.FloatProperty()
     date = ndb.IntegerProperty()
     user = ndb.StringProperty()
